api/requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
psycopg2-binary==2.9.9
pika==1.3.2
requests==2.32.3
beautifulsoup4==4.12.3
flask-cors==5.0.0
